♻️ Refactor quest category filter to component
