+++ src/routes/runs/[quest=leaderboard]/[category=category]/filters/DefaultRunFilter.svelte
@@ -1,34 +1,26 @@
 <script lang="ts">
 	import Divider from '$lib/Components/Divider.svelte';
 	import RunFilterModal from '../../../RunFilterModal.svelte';
-	import { NgsRunCategories } from '$lib/types/api/runs/categories';
 	import RunFilterTags from '../../../RunFilterTags.svelte';
 	import RunRules from '../../../RunRules.svelte';
 	import { runFilters } from '../../../runFilter';
-	import RadioOptions from '$lib/Components/RadioOptions.svelte';
 	import type { LeaderboardDefinition } from '$lib/leaderboard/leaderboard';
 	import { NgsQuests } from '$lib/types/api/runs/quests';
 	import type { ComponentType } from 'svelte';
 	import DfAegisSupportFilter from './DfAegisSupportFilter.svelte';
-	import { PartySize } from '$lib/types/api/partySizes';
 	import { goto } from '$app/navigation';
 	import { mapCategoryToRoute } from '../../../../../params/category';
 	import DuelDetailsFilter from './DuelDetailsFilter.svelte';
 	import QuestRankFilter from './QuestRankFilter.svelte';
 	import PartySizeFilter from './PartySizeFilter.svelte';
+	import QuestCategoryFilter from './QuestCategoryFilter.svelte';
+	import type { NgsRunCategories } from '$lib/types/api/runs/categories';
 
 	export let solo: boolean;
 	export let categories: NgsRunCategories[];
 	export let boardInfo: LeaderboardDefinition<any, any>;
 
 	$: $runFilters.rank = boardInfo.maxQuestRank.toString();
-
-	const categoryOptions = [
-		{ label: 'Quest', value: NgsRunCategories.Quest },
-		{ label: 'Urgent Quest', value: NgsRunCategories.UrgentQuest },
-		{ label: 'Trigger', value: NgsRunCategories.Trigger },
-	];
-	$: selectableCategories = categoryOptions.filter((c) => categories.includes(c.value));
 
 	interface FilterHandler {
 		filterComponent: ComponentType;
@@ -62,14 +54,7 @@
 >
 	<div class="flex flex-row flex-wrap place-content-center items-stretch gap-2">
 		<div class=" flex w-full gap-4 px-4 py-2 lg:gap-8">
-			{#if selectableCategories.length !== 0}
-				<RadioOptions
-					name="category"
-					options={selectableCategories}
-					on:changed={categoryChanged}
-					bind:value={$runFilters.category}
-				/>
-			{/if}
+			<QuestCategoryFilter {categories} on:categoryChanged={categoryChanged} />
 			{#if boardInfo.quest === NgsQuests.Duels}
 				<DuelDetailsFilter />
 			{/if}
